cyclic, cmd: qoute cyclic table name, fatal if loopback and fix a data race (#701)
